+++ tests/testthat/test-one_sample.R
@@ -144,6 +144,8 @@ test_that(
     # check Bayes factor values
     expect_equal(df_results$bf10[[1]], 5.958171e+20, tolerance = 0.001)
 
+    expect_snapshot(names(df_results))
+
     # extracting subtitle (without NA)
     set.seed(123)
     subtitle <-